+++ nix/launcher-config.nix
@@ -27,8 +27,8 @@ let
       cluster = "mainnet";
       networkName = "mainnet";
     };
-    shelley_testnet_v5 = {
-      cardanoEnv = cardanoLib.environments.mainnet_candidate_3;
+    shelley_testnet_v6 = {
+      cardanoEnv = cardanoLib.environments.mainnet_candidate_4;
       cluster = "shelley_testnet";
       networkName = "shelley_testnet";
     };
@@ -84,7 +84,7 @@ let
       staging = "Staging";
       testnet = "Testnet";
       shelley_testnet = "Shelley Testnet";
-      shelley_testnet_v5 = "Shelley Testnet v5";
+      shelley_testnet_v6 = "Shelley Testnet v6";
       shelley_qa = "Shelley QA";
     };
     unsupported = "Unsupported";