+++ source/renderer/app/components/staking/widgets/TooltipPool.js
@@ -62,7 +62,7 @@ const messages = defineMessages({
   producedBlocks: {
     id: 'staking.stakePools.tooltip.producedBlocks',
     defaultMessage: '!!!Produced blocks:',
-    description: '"Produced blocks" for the Stake Pools Tooltip page.',
+    description: '"Blocks" for the Stake Pools Tooltip page.',
   },
   retirement: {
     id: 'staking.stakePools.tooltip.retirement',