+++ source/renderer/app/domains/News.js
@@ -11,6 +11,8 @@ export type NewsAction = {
   event?: string,
 };
 
+export type IncidentColor = 'red' | 'theme-default' | 'grey';
+
 export const NewsTypes: {
   INCIDENT: NewsType,
   ALERT: NewsType,
@@ -21,6 +23,16 @@ export const NewsTypes: {
   ALERT: 'alert',
   ANNOUNCEMENT: 'announcement',
   INFO: 'info',
+};
+
+export const IncidentColors: {
+  RED: IncidentColor,
+  THEME_DEFAULT: IncidentColor,
+  GREY: IncidentColor,
+} = {
+  RED: 'red',
+  THEME_DEFAULT: 'theme-default',
+  GREY: 'grey',
 };
 
 export type NewsTypesStateType = {
@@ -40,6 +52,8 @@ class News {
   @observable date: number;
   @observable type: NewsType;
   @observable read: boolean;
+  @observable color: ?IncidentColor;
+  @observable repeatOnStartup: ?boolean;
 
   constructor(data: {
     id: number,
@@ -50,6 +64,8 @@ class News {
     date: number,
     type: NewsType,
     read: boolean,
+    color?: ?IncidentColor,
+    repeatOnStartup?: ?boolean,
   }) {
     Object.assign(this, data);
   }