+++ source/renderer/app/themes/daedalus/white.js
@@ -368,6 +368,42 @@ export const WHITE_THEME_OUTPUT = {
     '--theme-news-feed-icon-toggle-hover-background-color':
       'rgba(41, 181, 149, 0.1)',
     '--theme-news-feed-no-fetch-color': '#2d2d2d',
+    '--theme-news-feed-incident-overlay-background-color':
+      'rgba(255, 255, 255, 0.98)',
+    '--theme-news-feed-incident-overlay-text-color': '#2d2d2d',
+    '--theme-news-feed-incident-overlay-button-background':
+      'rgba(41, 181, 149, 0.1)',
+    '--theme-news-feed-incident-overlay-button-color-hover':
+      'rgba(255, 255, 255, 1)',
+    '--theme-news-feed-incident-overlay-button-color': '#29b595',
+    '--theme-news-feed-incident-overlay-scrollbar-thumb-background':
+      'rgba(45, 45, 45, 0.3)',
+    '--theme-news-feed-incident-overlay-scrollbar-thumb-background-hove':
+      'rgba(45, 45, 45, 0.5)',
+    '--theme-news-feed-incident-overlay-content-list-color':
+      'rgba(45, 45, 45, 0.7)',
+    '--theme-news-feed-incident-overlay-content-background':
+      'rgba(41, 181, 149, 0.1)',
+    '--theme-news-feed-incident-grey-overlay-background-color':
+      'rgba(68, 68, 68, 0.98)',
+    '--theme-news-feed-incident-grey-overlay-text-color': '#ffffff',
+    '--theme-news-feed-incident-grey-overlay-button-color': '#ffffff',
+    '--theme-news-feed-incident-grey-overlay-scrollbar-thumb-background':
+      'rgba(255, 255, 255, 0.3)',
+    '--theme-news-feed-incident-grey-overlay-scrollbar-thumb-background-hove':
+      'rgba(255, 255, 255, 0.5)',
+    '--theme-news-feed-incident-grey-overlay-content-list-color':
+      'rgba(255, 255, 255, 0.7)',
+    '--theme-news-feed-incident-red-overlay-background-color':
+      'rgba(171, 23, 0, 0.98)',
+    '--theme-news-feed-incident-red-overlay-text-color': '#ffffff',
+    '--theme-news-feed-incident-red-overlay-button-color': '#ffffff',
+    '--theme-news-feed-incident-red-overlay-scrollbar-thumb-background':
+      'rgba(255, 255, 255, 0.3)',
+    '--theme-news-feed-incident-red-overlay-scrollbar-thumb-background-hove':
+      'rgba(255, 255, 255, 0.5)',
+    '--theme-news-feed-incident-red-overlay-content-list-color':
+      'rgba(255, 255, 255, 0.7)',
   },
   newsItem: {
     '--theme-news-item-action-button-background-color':