+++ source/renderer/app/themes/utils/createTheme.js
@@ -659,6 +659,53 @@ export const createDaedalusComponentsTheme = (
       '--theme-news-feed-icon-toggle-hover-background-color':
         'rgba(0, 0, 0, 0.1)',
       '--theme-news-feed-no-fetch-color': '#fafbfc',
+      '--theme-news-feed-incident-overlay-background-color': `${chroma(
+        background.secondary.regular
+      ).alpha(0.96)}`,
+      '--theme-news-feed-incident-overlay-text-color': `${text.primary}`,
+      '--theme-news-feed-incident-overlay-button-color': `${text.primary}`,
+      '--theme-news-feed-incident-overlay-scrollbar-thumb-background': `${chroma(
+        text.primary
+      ).alpha(0.3)}`,
+      '--theme-news-feed-incident-overlay-scrollbar-thumb-background-hove': `${chroma(
+        text.primary
+      ).alpha(0.3)}`,
+      '--theme-news-feed-incident-overlay-content-list-color': `${chroma(
+        text.primary
+      ).alpha(0.7)}`,
+      '--theme-news-feed-incident-overlay-content-background':
+        'rgba(0, 0, 0, 0.1)',
+      '--theme-news-feed-incident-overlay-button-background':
+        'rgba(0, 0, 0, 0.1)',
+      '--theme-news-feed-incident-overlay-button-color-hover': `${chroma(
+        background.secondary.regular
+      ).alpha(0.96)}`,
+      '--theme-news-feed-incident-grey-overlay-background-color':
+        'rgba(68, 68, 68, 0.98)',
+      '--theme-news-feed-incident-grey-overlay-text-color': `${text.primary}`,
+      '--theme-news-feed-incident-grey-overlay-button-color': `${text.primary}`,
+      '--theme-news-feed-incident-grey-overlay-scrollbar-thumb-background': `${chroma(
+        text.primary
+      ).alpha(0.3)}`,
+      '--theme-news-feed-incident-grey-overlay-scrollbar-thumb-background-hove': `${chroma(
+        text.primary
+      ).alpha(0.5)}`,
+      '--theme-news-feed-incident-grey-overlay-content-list-color': `${chroma(
+        text.primary
+      ).alpha(0.7)}`,
+      '--theme-news-feed-incident-red-overlay-background-color':
+        'rgba(171, 23, 0, 0.98)',
+      '--theme-news-feed-incident-red-overlay-text-color': `${text.primary}`,
+      '--theme-news-feed-incident-red-overlay-button-color': `${text.primary}`,
+      '--theme-news-feed-incident-red-overlay-scrollbar-thumb-background': `${chroma(
+        text.primary
+      ).alpha(0.3)}`,
+      '--theme-news-feed-incident-red-overlay-scrollbar-thumb-background-hove': `${chroma(
+        text.primary
+      ).alpha(0.5)}`,
+      '--theme-news-feed-incident-red-overlay-content-list-color': `${chroma(
+        text.primary
+      ).alpha(0.7)}`,
     },
     newsItem: {
       '--theme-news-item-action-button-background-color': 'rgba(0, 0, 0, 0.1)',